resolved constant sidebar issue
